Updated some legacy code that broke unit tests under Java13 in VGL. Also refactored some test code that was in main into test directory.
