+++ pom.xml
@@ -459,21 +459,12 @@
             	</exclusion>
             </exclusions>
         </dependency>
-        <dependency>
-            <groupId>org.jmock</groupId>
-            <artifactId>jmock-legacy</artifactId>
-            <version>2.12.0</version>
-            <exclusions>
-            	<exclusion>
-            		<groupId>org.ow2.asm</groupId>
-            		<artifactId>asm</artifactId>
-            	</exclusion>
-            	<exclusion>
-            		<groupId>cglib</groupId>
-            		<artifactId>cglib</artifactId>
-            	</exclusion>
-            </exclusions>
-        </dependency>
+		<dependency>
+		    <groupId>org.jmock</groupId>
+		    <artifactId>jmock-imposters</artifactId>
+		    <version>2.12.0</version>
+		    <scope>test</scope>
+		</dependency>
    		<dependency>
    			<groupId>cglib</groupId>
    			<artifactId>cglib</artifactId>
