+++ src/main/java/org/auscope/portal/core/util/ResourceUtil.java
@@ -1,9 +1,7 @@
-package org.auscope.portal.core.test;
+package org.auscope.portal.core.util;
 
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 
 import org.apache.commons.io.IOUtils;
