+++ src/test/java/org/auscope/portal/core/server/controllers/TestGetCapabilitiesController.java
@@ -4,14 +4,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 
-import org.auscope.portal.core.server.controllers.WMSController;
 import org.auscope.portal.core.services.WMSService;
 import org.auscope.portal.core.services.responses.csw.CSWRecord;
 import org.auscope.portal.core.services.responses.wms.GetCapabilitiesRecord;
 import org.auscope.portal.core.services.responses.wms.GetCapabilitiesRecord_1_1_1;
 import org.auscope.portal.core.services.responses.wms.GetCapabilitiesRecord_1_3_0;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.auscope.portal.core.view.ViewCSWRecordFactory;
 import org.auscope.portal.core.view.ViewKnownLayerFactory;
 import org.jmock.Expectations;
