+++ src/test/java/org/auscope/portal/core/services/TestNagios4Service.java
@@ -11,7 +11,7 @@ import org.auscope.portal.core.services.methodmakers.Nagios4MethodMaker;
 import org.auscope.portal.core.services.responses.nagios.ServiceStatusResponse;
 import org.auscope.portal.core.services.responses.nagios.ServiceStatusResponse.Status;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.jmock.Expectations;
 import org.junit.Assert;
 import org.junit.Before;
