+++ src/test/java/org/auscope/portal/core/services/TestNamespaceService.java
@@ -1,17 +1,17 @@
 package org.auscope.portal.core.services;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
+
 import org.apache.http.client.methods.HttpRequestBase;
 import org.auscope.portal.core.server.http.HttpServiceCaller;
 import org.auscope.portal.core.services.methodmakers.WFSGetFeatureMethodMaker;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.jmock.Expectations;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-
-import java.io.IOException;
-import java.net.URISyntaxException;
 
 public class TestNamespaceService extends PortalTestClass {
