+++ src/test/java/org/auscope/portal/core/services/TestSISSVoc2Service.java
@@ -10,7 +10,7 @@ import org.auscope.portal.core.services.methodmakers.sissvoc.SISSVoc2MethodMaker
 import org.auscope.portal.core.services.responses.vocab.Concept;
 import org.auscope.portal.core.services.responses.vocab.ConceptFactory;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.jmock.Expectations;
 import org.junit.Assert;
 import org.junit.Before;
