+++ src/test/java/org/auscope/portal/core/services/responses/csw/TestCSWGetRecordResponse.java
@@ -8,8 +8,8 @@ import javax.xml.xpath.XPathExpressionException;
 
 import org.auscope.portal.core.services.csw.CSWServiceItem;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
 import org.auscope.portal.core.util.DOMUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
