+++ src/test/java/org/auscope/portal/core/services/responses/wcs/TestDescribeCoverageRecord.java
@@ -10,8 +10,8 @@ import javax.xml.xpath.XPathExpressionException;
 
 import org.auscope.portal.core.services.responses.ows.OWSException;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
 import org.auscope.portal.core.util.DOMUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.w3c.dom.DOMException;
