+++ src/test/java/org/auscope/portal/core/services/responses/wms/TestGetCapabilitiesRecord.java
@@ -8,7 +8,7 @@ import javax.xml.xpath.XPathExpressionException;
 
 import org.auscope.portal.core.services.responses.csw.CSWGeographicBoundingBox;
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
+import org.auscope.portal.core.util.ResourceUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.xml.sax.SAXException;
