+++ src/test/java/org/auscope/portal/core/test/AbstractFilterTestUtilities.java
