+++ src/test/java/org/auscope/portal/core/test/AssertViewUtility.java
