+++ src/test/java/org/auscope/portal/core/test/jmock/FileWithNameMatcher.java
