+++ src/test/java/org/auscope/portal/core/util/TestDOMUtil.java
@@ -16,7 +16,6 @@ import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
 
 import org.auscope.portal.core.test.PortalTestClass;
-import org.auscope.portal.core.test.ResourceUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.w3c.dom.Document;
